T: improve tests for RsSmartEnterProcessor
* Perform `EditorCompleteStatement` action instead of smart enter processors explicit call
* Inline texts of test cases and drop test data files
