+++ src/test/kotlin/org/rust/ide/typing/assist/RsSmartEnterProcessorTest.kt
@@ -5,42 +5,146 @@
 
 package org.rust.ide.typing.assist
 
-import com.intellij.codeInsight.editorActions.smartEnter.SmartEnterProcessors
-import com.intellij.lang.Language
-import com.intellij.openapi.command.WriteCommandAction
+import com.intellij.openapi.actionSystem.IdeActions
+import org.intellij.lang.annotations.Language
 import org.rust.RsTestBase
-import org.rust.lang.RsLanguage
 
 /**
  * Unit tests for [RsSmartEnterProcessor]
  */
 class RsSmartEnterProcessorTest : RsTestBase() {
 
-    override val dataPath = "org/rust/ide/typing/assist/fixtures"
-
-    fun getSmartProcessors(language: Language) = SmartEnterProcessors.INSTANCE.forKey(language)
-
-    private fun doTest() {
-        myFixture.configureByFile(fileName)
-        val processors = getSmartProcessors(RsLanguage)
-
-        WriteCommandAction.writeCommandAction(project).run<Throwable> {
-            val editor = myFixture.editor
-            for (processor in processors) {
-                processor.process(project, editor, myFixture.file)
-            }
+    fun `test fix simple method call`() = doTest("""
+        fn f() -> i32 {
+            /*caret*/f(
         }
+    """, """
+        fn f() -> i32 {
+            f();
+            /*caret*/
+        }
+    """)
+    fun `test fix nested method call`() = doTest("""
+        fn double(x: i32) -> i32 {
+        /*caret*/double(double(x
+            double(x)
+        }
+    """, """
+        fn double(x: i32) -> i32 {
+            double(double(x));
+            /*caret*/
+            double(x)
+        }
+    """)
+    
+    fun `test fix method call with string literal`() = doTest("""
+        fn f(s: String) -> String {
+            f(f(f("((")/*caret*/
+        }
+    """, """
+        fn f(s: String) -> String {
+            f(f(f("((")));
+            /*caret*/
+        }
+    """)
+    
+    fun `test fix method call multiple lines`() = doTest("""
+        fn f(s: String) -> String {
+            f("");
+            f(
+                f("(("/*caret*/
+        }
+    """, """
+        fn f(s: String) -> String {
+            f("");
+            f(
+                f("(("));
+            /*caret*/
+        }
+    """)
+    
+    fun `test fix whitespace and semicolon`() = doTest("""
+        fn f(x: i32) -> i32 {
+            f(f(x))/*caret*/  ;
+        }
+    """, """
+        fn f(x: i32) -> i32 {
+            f(f(x));
+            /*caret*/
+        }
+    """)
+    
+    fun `test fix semicolon after declaration`() = doTest("""
+        struct Point {
+            x: i32,
+            y: i32,
+        }
+        
+        fn main() {
+            let origin = Point { x: 0, y: 0 }/*caret*/
+        }
+    """, """
+        struct Point {
+            x: i32,
+            y: i32,
+        }
+        
+        fn main() {
+            let origin = Point { x: 0, y: 0 };
+            /*caret*/
+        }
+    """)
+    
+    fun `test fix declaration with call`() = doTest("""
+        fn f() -> i32 {
+            return 42;
+        }
+        
+        fn main() {
+            let x = f(/*caret*/
+        }
+    """, """
+        fn f() -> i32 {
+            return 42;
+        }
+        
+        fn main() {
+            let x = f();
+            /*caret*/
+        }
+    """)
+    
+    fun `test fix match in let`() = doTest("""
+        fn main() {
+            let version_req = match version {
+                Some(v) => try!(VersionReq::parse(v)),
+                None => VersionReq::any()
+            }/*caret*/
+        }
+    """, """
+        fn main() {
+            let version_req = match version {
+                Some(v) => try!(VersionReq::parse(v)),
+                None => VersionReq::any()
+            };
+            /*caret*/
+        }
+    """)
+    
+    fun `test fix call in stmt`() = doTest("""
+        fn f(s: String) {
+            /*caret*/f(
+            let x = 5;
+        }
+    """, """
+        fn f(s: String) {
+            f();
+            /*caret*/
+            let x = 5;
+        }
+    """)
 
-        myFixture.checkResultByFile(fileName.replace(".rs", "_after.rs"), true)
+    private fun doTest(@Language("Rust") before: String, @Language("Rust") after: String) = checkByText(before, after) {
+        myFixture.performEditorAction(IdeActions.ACTION_EDITOR_COMPLETE_STATEMENT)
     }
-
-    fun `test fix simple method call`() = doTest()
-    fun `test fix nested method call`() = doTest()
-    fun `test fix method call with string literal`() = doTest()
-    fun `test fix method call multiple lines`() = doTest()
-    fun `test fix whitespace and semicolon`() = doTest()
-    fun `test fix semicolon after declaration`() = doTest()
-    fun `test fix declaration with call`() = doTest()
-    fun `test fix match in let`() = doTest()
-    fun `test fix call in stmt`() = doTest()
 }
