--- src/test/resources/org/rust/ide/typing/assist/fixtures/fix_call_in_stmt.rs
@@ -1,4 +1,0 @@
-fn f(s: String) {
-    <caret>f(
-    let x = 5;
-}
