--- src/test/resources/org/rust/ide/typing/assist/fixtures/fix_call_in_stmt_after.rs
@@ -1,5 +1,0 @@
-fn f(s: String) {
-    f();
-
-    let x = 5;
-}
