--- src/test/resources/org/rust/ide/typing/assist/fixtures/fix_declaration_with_call.rs
@@ -1,7 +1,0 @@
-fn f() -> i32 {
-    return 42;
-}
-
-fn main() {
-    let x = f(<caret>
-}
