--- src/test/resources/org/rust/ide/typing/assist/fixtures/fix_declaration_with_call_after.rs
@@ -1,8 +1,0 @@
-fn f() -> i32 {
-    return 42;
-}
-
-fn main() {
-    let x = f();
-    <caret>
-}
