--- src/test/resources/org/rust/ide/typing/assist/fixtures/fix_match_in_let.rs
@@ -1,6 +1,0 @@
-fn main() {
-    let version_req = match version {
-        Some(v) => try!(VersionReq::parse(v)),
-        None => VersionReq::any()
-    }<caret>
-}
