--- src/test/resources/org/rust/ide/typing/assist/fixtures/fix_method_call_multiple_lines_after.rs
@@ -1,6 +1,0 @@
-fn f(s: String) -> String {
-    f("");
-    f(
-        f("(("));
-    <caret>
-}
