--- src/test/resources/org/rust/ide/typing/assist/fixtures/fix_method_call_with_string_literal.rs
@@ -1,3 +1,0 @@
-fn f(s: String) -> String {
-    f(f(f("((")<caret>
-}
