--- src/test/resources/org/rust/ide/typing/assist/fixtures/fix_nested_method_call.rs
@@ -1,4 +1,0 @@
-fn double(x: i32) -> i32 {
-<caret>double(double(x
-    double(x)
-}
