--- src/test/resources/org/rust/ide/typing/assist/fixtures/fix_nested_method_call_after.rs
@@ -1,5 +1,0 @@
-fn double(x: i32) -> i32 {
-    double(double(x));
-    <caret>
-    double(x)
-}
