--- src/test/resources/org/rust/ide/typing/assist/fixtures/fix_semicolon_after_declaration.rs
@@ -1,8 +1,0 @@
-struct Point {
-    x: i32,
-    y: i32,
-}
-
-fn main() {
-    let origin = Point { x: 0, y: 0 }<caret>
-}
