--- src/test/resources/org/rust/ide/typing/assist/fixtures/fix_simple_method_call.rs
@@ -1,3 +1,0 @@
-fn f() -> i32 {
-    <caret>f(
-}
