--- src/test/resources/org/rust/ide/typing/assist/fixtures/fix_simple_method_call_after.rs
@@ -1,4 +1,0 @@
-fn f() -> i32 {
-    f();
-    <caret>
-}
