--- src/test/resources/org/rust/ide/typing/assist/fixtures/fix_whitespace_and_semicolon_after.rs
@@ -1,4 +1,0 @@
-fn f(x: i32) -> i32 {
-    f(f(x));
-    <caret>
-}
